+++ src/store/types/TimeType.ts
@@ -1,0 +1,4 @@
+export type Time = {
+    timezone: number 
+    }
+
