+++ src/store/types/WeatherType.ts
@@ -1,0 +1,14 @@
+export type Weather = {
+    main:  {
+        temp: number,
+        feels_like: number,
+        pressure: number
+    },
+    wind: {
+        speed: number,
+    }
+    ,
+    clouds: {
+        all : number,
+    }
+}
