+++ src/store/types/WeekType.ts
@@ -1,0 +1,18 @@
+export interface Week {
+    list: {
+      dt: number,
+      main: {
+        temp: number;
+        feels_like: number;
+      },
+      weather:{
+        description: string,
+      }[],
+      
+      }[];
+    isLoading: boolean;
+    response: {
+      status: number;
+      message: string;
+    };}
+
